Preserve Annotated metadata in modules pipeline
