+++ macrotype/modules/emit.py
@@ -226,7 +226,7 @@ def stringify_annotation(ann: Any, name_map: dict[int, str]) -> str:
         first, *metas = args
         parts = [stringify_annotation(first, name_map)]
         for meta in metas:
-            parts.append(name_map.get(id(meta), _qualname(meta)))
+            parts.append(_qualname(meta))
         return f"Annotated[{', '.join(parts)}]"
 
     if origin is tuple and ann is not tuple and not args:
